refactor(hickey): consolidate applyHighlights into core
Hickey #3: iframe/index.ts and highlightOverlay.ts both implemented the
same six-step pipeline (guard on Highlight + CSS.highlights → extract
text → loop comments → findQuote → rangeFromOffsets → Highlight.set).
The only differences were the root node and the highlight name —
exactly the parameters a shared function takes. Extract to
`packages/artifact-sdk/src/core/applyHighlights.ts`; both call sites
delegate. Drift between the two surfaces is now impossible.
