+++ packages/artifact-sdk/src/core/index.ts
@@ -1,6 +1,10 @@
 export { extractQuote } from "./extractQuote";
 export { findQuote, rangeFromOffsets, type QuoteMatch } from "./findQuote";
 export {
+  applyHighlights,
+  type HighlightInputComment,
+} from "./applyHighlights";
+export {
   COMMENT_HIGHLIGHT_BACKGROUND,
   COMMENT_HIGHLIGHT_UNDERLINE,
   COMMENT_HIGHLIGHT_STYLE,
